Split single point read-write test functions
