+++ cql_storage_types/cql_storage_type_testing_lib/Cargo.toml
@@ -4,7 +4,7 @@ description = "Testing library for CQL Database storage types"
 readme = "./README.md"
 keywords = ["cql"]
 categories = ["development-tools::testing"]
-version = "0.3.0"
+version = "0.4.0"
 repository = "https://github.com/AndrewSisley/CQLDb/tree/master/cql_storage_types/cql_storage_type_testing_lib"
 authors = ["Andrew Sisley"]
 edition = "2018"
